+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.2.3 (2016-01-29)
 - Introduce ```sourceFiles``` global setting (used by eslint, jscs and check-file-names)
 - Add JSON loader to the default webpack config
 - Enhance error logging of webpack task